src/requirements.txt
spotipy
PyYAML
